+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+0.9.1 (2019-03-?)
+----------------
+* Fixed retrieval of added and removed lines when there are spaces in a file name.
+* Fixed indexed input in `get_mass_changes`
+
 0.9 (2019-03-19)
 ----------------
 
@@ -11,7 +16,7 @@ History
 0.8.2 (2019-02-26)
 ------------------
 
-* Added svn.get_diff_stats to retrieve line changes stats per diff.
+* Added `svn.get_diff_stats` to retrieve line changes stats per diff.
 
 0.8 (2019-02-13)
 ----------------